moved semaphore into a separate header file.
